Added file with documentation of the vocabulary.
The file has the format of the Emacs Org mode so that it can be relatively
easily converted to a proper documentation format, once we have found one.

Most kernel words have a short explanation, except some dictionary-related
words: We will have to think about them a bit and change them, perhaps.